+++ src/main/java/net/frozenblock/lib/debug/client/impl/DebugRenderManager.java
@@ -18,30 +18,103 @@
 package net.frozenblock.lib.debug.client.impl;
 
 import it.unimi.dsi.fastutil.objects.Object2ObjectLinkedOpenHashMap;
+import java.util.ArrayList;
+import java.util.Comparator;
+import java.util.List;
+import java.util.Map;
+import java.util.stream.Stream;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
 import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientTickEvents;
-import net.frozenblock.lib.debug.client.ImprovedGameEventListenerRenderer;
-import net.frozenblock.lib.debug.client.ImprovedGoalSelectorDebugRenderer;
+import net.fabricmc.fabric.api.client.networking.v1.ClientPlayNetworking;
+import net.frozenblock.lib.FrozenSharedConstants;
+import net.frozenblock.lib.debug.client.api.DebugRendererEvents;
+import net.frozenblock.lib.debug.client.renderer.ImprovedGameEventListenerRenderer;
+import net.frozenblock.lib.debug.client.renderer.ImprovedGoalSelectorDebugRenderer;
+import net.frozenblock.lib.debug.networking.GoalDebugRemovePayload;
+import net.frozenblock.lib.debug.networking.ImprovedGameEventDebugPayload;
+import net.frozenblock.lib.debug.networking.ImprovedGameEventListenerDebugPayload;
+import net.frozenblock.lib.debug.networking.ImprovedGoalDebugPayload;
+import net.minecraft.client.Minecraft;
 import net.minecraft.resources.ResourceLocation;
-import java.util.Map;
+import net.minecraft.world.entity.Entity;
 
 @Environment(EnvType.CLIENT)
 public class DebugRenderManager {
-	public static final Map<DebugRendererHolder, ResourceLocation> DEBUG_RENDERER_HOLDERS = new Object2ObjectLinkedOpenHashMap<>();
-
-	public static ImprovedGoalSelectorDebugRenderer improvedGoalSelectorRenderer;
-	public static ImprovedGameEventListenerRenderer improvedGameEventListenerRenderer;
+	public static Map<DebugRendererHolder, ResourceLocation> DEBUG_RENDERER_HOLDERS = new Object2ObjectLinkedOpenHashMap<>();
+	private static final List<Runnable> ON_CLEAR_RUNNABLES = new ArrayList<>();
 
 	public static void registerRenderer(ResourceLocation location, DebugRendererHolder.RenderInterface renderInterface) {
+		if (location == null) throw new IllegalArgumentException("ResourceLocation cannot be null!");
 		DEBUG_RENDERER_HOLDERS.put(new DebugRendererHolder(renderInterface), location);
+		Stream<Map.Entry<DebugRendererHolder, ResourceLocation>> entries = DEBUG_RENDERER_HOLDERS.entrySet().stream()
+			.sorted(Comparator.comparing(value ->  value.getValue().getPath()));
+
+		Map<DebugRendererHolder, ResourceLocation> newRenderHolders = new Object2ObjectLinkedOpenHashMap<>();
+		entries.forEach(entry -> newRenderHolders.put(entry.getKey(), entry.getValue()));
+
+		DEBUG_RENDERER_HOLDERS = newRenderHolders;
 	}
 
 	public static void init() {
-		ClientTickEvents.START_WORLD_TICK.register(clientLevel -> {
-			if (improvedGameEventListenerRenderer != null) {
-				improvedGameEventListenerRenderer.tick();
-			}
+		DebugRendererEvents.DEBUG_RENDERERS_CREATED.register(client -> {
+			ImprovedGoalSelectorDebugRenderer improvedGoalSelectorRenderer = new ImprovedGoalSelectorDebugRenderer(client);
+
+			ClientPlayNetworking.registerGlobalReceiver(ImprovedGoalDebugPayload.PACKET_TYPE, (packet, ctx) -> {
+				Entity entity = ctx.client().level.getEntity(packet.entityId());
+				if (entity != null) {
+					improvedGoalSelectorRenderer.addGoalSelector(
+						entity,
+						packet.goals()
+					);
+				}
+			});
+
+			ClientPlayNetworking.registerGlobalReceiver(GoalDebugRemovePayload.PACKET_TYPE, (packet, ctx) -> {
+				improvedGoalSelectorRenderer.removeGoalSelector(packet.entityId());
+			});
+
+			addClearRunnable(improvedGoalSelectorRenderer::clear);
+
+			registerRenderer(FrozenSharedConstants.id("goal"), improvedGoalSelectorRenderer::render);
 		});
+
+		DebugRendererEvents.DEBUG_RENDERERS_CREATED.register(client -> {
+			ImprovedGameEventListenerRenderer improvedGameEventRenderer = new ImprovedGameEventListenerRenderer(client);
+
+			ClientPlayNetworking.registerGlobalReceiver(ImprovedGameEventListenerDebugPayload.PACKET_TYPE, (packet, ctx) -> {
+				improvedGameEventRenderer.trackListener(
+					packet.listenerPos(),
+					packet.listenerRange()
+				);
+			});
+
+			ClientPlayNetworking.registerGlobalReceiver(ImprovedGameEventDebugPayload.PACKET_TYPE, (packet, ctx) -> {
+				improvedGameEventRenderer.trackGameEvent(
+					packet.gameEventType(),
+					packet.pos()
+				);
+			});
+
+			ClientTickEvents.START_WORLD_TICK.register(clientLevel -> {
+				improvedGameEventRenderer.tick();
+			});
+
+			registerRenderer(FrozenSharedConstants.id("game_event"), improvedGameEventRenderer::render);
+		});
+	}
+
+	public static void clearAdditionalRenderers() {
+		ON_CLEAR_RUNNABLES.forEach(Runnable::run);
+	}
+
+	public static void addClearRunnable(Runnable runnable) {
+		ON_CLEAR_RUNNABLES.add(runnable);
+	}
+
+	public static float PARTIAL_TICK;
+
+	public static void updatePartialTick() {
+		PARTIAL_TICK = Minecraft.getInstance().getTimer().getGameTimeDeltaPartialTick(true);
 	}
 }