+++ src/main/java/net/frozenblock/lib/debug/mixin/PathMixin.java
@@ -19,6 +19,7 @@ package net.frozenblock.lib.debug.mixin;
 
 import java.util.List;
 import java.util.Set;
+import net.frozenblock.lib.config.frozenlib_config.FrozenLibConfig;
 import net.minecraft.core.BlockPos;
 import net.minecraft.network.FriendlyByteBuf;
 import net.minecraft.world.level.pathfinder.Node;
@@ -42,17 +43,19 @@ public class PathMixin {
 	private BlockPos target;
 
 	@Inject(method = "writeToStream", at = @At("HEAD"))
-	private void frozenLib$writeToStream(FriendlyByteBuf buf, CallbackInfo ci) {
-		this.debugData = new Path.DebugData(
-			this.nodes.stream().filter((pathNode) -> !pathNode.closed).toArray(Node[]::new),
-			this.nodes.stream().filter((pathNode) -> pathNode.closed).toArray(Node[]::new),
-			Set.of(
-				new Target(
-					this.target.getX(),
-					this.target.getY(),
-					this.target.getZ()
+	private void frozenLib$writeToStream(FriendlyByteBuf buf, CallbackInfo info) {
+		if (FrozenLibConfig.IS_DEBUG) {
+			this.debugData = new Path.DebugData(
+				this.nodes.stream().filter((pathNode) -> !pathNode.closed).toArray(Node[]::new),
+				this.nodes.stream().filter((pathNode) -> pathNode.closed).toArray(Node[]::new),
+				Set.of(
+					new Target(
+						this.target.getX(),
+						this.target.getY(),
+						this.target.getZ()
+					)
 				)
-			)
-		);
+			);
+		}
 	}
 }