+++ CHANGELOG.md
@@ -5,3 +5,7 @@ Make sure to clear this after each release
 Put changelog here:
 
 -----------------
+- Implemented debug renderers while in a development environment.
+  - Requires that "isDebug" in FrozenLib's config is set to true.
+  - Will add a new button on the pause menu to toggle different debug renderers.
+- Added an API and event to easily register new debug renderers.